feat: implement external supplier inventory purchase workflow

+++ src/main/java/com/saikiran/inventory/product/dto/request/ProductVariantIdRequest.java
@@ -20,7 +20,7 @@ public class ProductVariantIdRequest {
 
     private UnitType unitType;
 
-    private int unitValue;
+    private BigDecimal unitValue;
 
     private BigDecimal currentPrice;
